Move confirmUserExists to validation.ts
